+++ test/integration/collection.test.ts
@@ -42,7 +42,7 @@ describe('Collection', () => {
         // todo: change here and below to raw webdriver assert implementation
         expect(await collection.matching(have.texts('John', 'Bob', 'Kate'))).toBe(true);
     });
-    
+
     it('is performed on each subsequent "ask"', async () => {
         await GIVEN.openedEmptyPage();
         const collection = browser.all('li');